✨ feat(basic-crud-service): switching the name of the app
