+++ apps/basic-crud-service/src/basic-crud-service.controller.ts
@@ -1,0 +1,12 @@
+import { Controller, Get } from '@nestjs/common';
+import { BasicCrudServiceService } from './basic-crud-service.service';
+
+@Controller()
+export class BasicCrudServiceController {
+  constructor(private readonly basicCrudServiceService: BasicCrudServiceService) {}
+
+  @Get()
+  getHello(): string {
+    return this.basicCrudServiceService.getHello();
+  }
+}
