+++ apps/basic-crud-service/src/basic-crud-service.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 
 @Injectable()
 export class BasicCrudServiceService {
-	getHello(): string {
-		return 'Hello World!';
-	}
+  getHello(): string {
+    return 'Hello World!';
+  }
 }
