+++ apps/basic-crud-service/test/jest-e2e.json
