--- apps/basic_crud_service/src/basic_crud_service.controller.ts
@@ -1,14 +1,0 @@
-import { Controller, Get } from '@nestjs/common';
-import { BasicCrudServiceService } from './basic_crud_service.service';
-
-@Controller()
-export class BasicCrudServiceController {
-	constructor(
-		private readonly basicCrudServiceService: BasicCrudServiceService,
-	) {}
-
-	@Get()
-	getHello(): string {
-		return this.basicCrudServiceService.getHello();
-	}
-}
