+++ package.json
@@ -1,79 +1,79 @@
 {
-	"name": "nest.js",
-	"version": "0.0.1",
-	"description": "",
-	"author": "Caio Maciel",
-	"private": true,
-	"license": "MIT",
-	"scripts": {
-		"build": "nest build",
-		"format": "prettier --write \"apps/**/*.ts\" \"libs/**/*.ts\"",
-		"start": "nest start",
-		"start:dev": "nest start --watch",
-		"start:debug": "nest start --debug --watch",
-		"start:prod": "node dist/apps/nest.js/main",
-		"lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
-		"test": "jest",
-		"test:watch": "jest --watch",
-		"test:cov": "jest --coverage",
-		"test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
-		"test:e2e": "jest --config ./apps/nest.js/test/jest-e2e.json",
-		"commit": "npx git-cz"
-	},
-	"dependencies": {
-		"@nestjs/common": "^10.0.0",
-		"@nestjs/core": "^10.0.0",
-		"@nestjs/platform-express": "^10.0.0",
-		"cz-emoji-conventional": "^1.0.2",
-		"reflect-metadata": "^0.1.13",
-		"rxjs": "^7.8.1"
-	},
-	"devDependencies": {
-		"@nestjs/cli": "^10.0.0",
-		"@nestjs/schematics": "^10.0.0",
-		"@nestjs/testing": "^10.0.0",
-		"@types/express": "^4.17.17",
-		"@types/jest": "^29.5.2",
-		"@types/node": "^20.3.1",
-		"@types/supertest": "^2.0.12",
-		"@typescript-eslint/eslint-plugin": "^6.0.0",
-		"@typescript-eslint/parser": "^6.0.0",
-		"eslint": "^8.42.0",
-		"eslint-config-prettier": "^9.0.0",
-		"eslint-plugin-prettier": "^5.0.0",
-		"jest": "^29.5.0",
-		"prettier": "^3.0.0",
-		"source-map-support": "^0.5.21",
-		"supertest": "^6.3.3",
-		"ts-jest": "^29.1.0",
-		"ts-loader": "^9.4.3",
-		"ts-node": "^10.9.1",
-		"tsconfig-paths": "^4.2.0",
-		"typescript": "^5.1.3"
-	},
-	"jest": {
-		"moduleFileExtensions": [
-			"js",
-			"json",
-			"ts"
-		],
-		"rootDir": ".",
-		"testRegex": ".*\\.spec\\.ts$",
-		"transform": {
-			"^.+\\.(t|j)s$": "ts-jest"
-		},
-		"collectCoverageFrom": [
-			"**/*.(t|j)s"
-		],
-		"coverageDirectory": "./coverage",
-		"testEnvironment": "node",
-		"roots": [
-			"<rootDir>/apps/"
-		]
-	},
-	"config": {
-		"commitizen": {
-			"path": "./node_modules/cz-emoji-conventional"
-		}
-	}
+  "name": "nest.js",
+  "version": "0.0.1",
+  "description": "",
+  "author": "Caio Maciel",
+  "private": true,
+  "license": "MIT",
+  "scripts": {
+    "build": "nest build",
+    "format": "prettier --write \"apps/**/*.ts\" \"libs/**/*.ts\"",
+    "start": "nest start",
+    "start:dev": "nest start --watch",
+    "start:debug": "nest start --debug --watch",
+    "start:prod": "node dist/apps/nest.js/main",
+    "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
+    "test": "jest",
+    "test:watch": "jest --watch",
+    "test:cov": "jest --coverage",
+    "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
+    "test:e2e": "jest --config ./apps/nest.js/test/jest-e2e.json",
+    "commit": "npx git-cz"
+  },
+  "dependencies": {
+    "@nestjs/common": "^10.0.0",
+    "@nestjs/core": "^10.0.0",
+    "@nestjs/platform-express": "^10.0.0",
+    "cz-emoji-conventional": "^1.0.2",
+    "reflect-metadata": "^0.1.13",
+    "rxjs": "^7.8.1"
+  },
+  "devDependencies": {
+    "@nestjs/cli": "^10.0.0",
+    "@nestjs/schematics": "^10.0.0",
+    "@nestjs/testing": "^10.0.0",
+    "@types/express": "^4.17.17",
+    "@types/jest": "^29.5.2",
+    "@types/node": "^20.3.1",
+    "@types/supertest": "^2.0.12",
+    "@typescript-eslint/eslint-plugin": "^6.0.0",
+    "@typescript-eslint/parser": "^6.0.0",
+    "eslint": "^8.42.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-prettier": "^5.0.0",
+    "jest": "^29.5.0",
+    "prettier": "^3.0.0",
+    "source-map-support": "^0.5.21",
+    "supertest": "^6.3.3",
+    "ts-jest": "^29.1.0",
+    "ts-loader": "^9.4.3",
+    "ts-node": "^10.9.1",
+    "tsconfig-paths": "^4.2.0",
+    "typescript": "^5.1.3"
+  },
+  "jest": {
+    "moduleFileExtensions": [
+      "js",
+      "json",
+      "ts"
+    ],
+    "rootDir": ".",
+    "testRegex": ".*\\.spec\\.ts$",
+    "transform": {
+      "^.+\\.(t|j)s$": "ts-jest"
+    },
+    "collectCoverageFrom": [
+      "**/*.(t|j)s"
+    ],
+    "coverageDirectory": "./coverage",
+    "testEnvironment": "node",
+    "roots": [
+      "<rootDir>/apps/"
+    ]
+  },
+  "config": {
+    "commitizen": {
+      "path": "./node_modules/cz-emoji-conventional"
+    }
+  }
 }
